:card_file_box: Initial intake catalog for ICESat-2 ATL06 data
Making sense of the ICESat-2 ATL06 directory structure through a intake YAML catalog. Uses a path_as_pattern regex string to break down the long filepath into its constituent meanings, and the file will also be cached on first download. The HDF (.h5) files are read using xarray's NetCDF driver, and we've got some hardcoded (but changeable) variables for the date (up to 2019.06.26, right before the safe-hold period from 2019.06.26 to 2019.07.09), orbital segment (10,11,12 are specific to Antarctica) and laser (gt2l is the center 'left' beam out of the 6). There's also a hvplot quickview that can be used to plot a sample of the points (using default settings) on a bokeh map.

The Python code to actually use this catalog will follow later. Note that ERS authentication is needed to download data mentioned in this catalog from NSIDC! Just wanted to commit this >2 month old work into git. Been messing around too much with OPeNDAP, Directory caching, and other things.