:construction: Shift PyViz dependencies from pypi to conda-forge
A work in progress on migrating a whole bunch of PyViz dependencies (from pypi to conda-forge), in an attempt to resolve issues with interactivity. The hvplot->holoviews->bokeh stack, intermingled with datashader, and geoviews, xrviz et al isn't playing nicely together when pip installed. Like, they sneakily install fine, but the interactive plotting (which is what they're there for!) isn't working as expected. Also unpinned pyproj, toolz and tqdm as they're installed in conda already, trying to reduce deduplication still.