:recycle: Turn BBox into a dataclass with bounds and subset methods
Refactor the BBox namedtuple into a Python 3.7+ dataclass, complete with convenience functions to get the bounding box coordinates as a tuple (either in a PyGMT or Shapely compatible format), as well as geographically subsetting an xarray.Dataset! This BBox class also allows us to set a name, and has a automatic 'scale' property, handy for making PyGMT plots later. Also moved the optional geographic subsetting section nearer to the top of the notebook as some people might want to do that before all the compute heavy stuff.