:speech_balloon: Update project description and pip install instructions
Not sure if I have time to make it deep learning, but machine learning is definitely possible. Also make sure we tell people how to `pip install deepicedrain` from Github! Added license badge, and included a teaser image, because a picture says a thousand words (and requires more than a thousand lines of code to produce). Took the opportunity to remove a few hundred lines of old code too!