:package: Flatten the ATL11z123 Zarr store directory structure
Not really a point in having a 'pt123' group in the ATL11*.zarr files, if it's the only one inside! Better to just open the Zarr folder and see all the variables directly. This is really just a cosmetic change, shaving out a line of code in some places. Bumped up pytx from 2019.3 to 2020.1 in order for tests to pass, will submit a proper fix after this.