:sparkles: Calculate height range using nanptp
Finding change where it's at, and quickly! There's nothing really special about calculating height range, it's just subtracting minimum height from maximum height. The `nanptp` function (which numpy doesn't have) merely accounts for NaN values, and I've put in the extra effort to parallelize it on dask using xr.apply_ufunc, so that it takes minutes to run on ~1 trillion points. Took a long detour packaging up `deepicedrain` properly before this dhdt notebook can be released, but it's worth it to polish out most of the cruft and reduce the amount of boilerplate preprocessing code. Also remove the need to close the atl11 test dataset as using to_dask() instead of read() seems to be nicer.