:recycle: Refactor data type downcasting to do so via encoding argument
The downcasted data types still weren't being written to Zarr properly (some data variables are still written as int64/float64). Worse still, I can't reproduce it on a MWE to file a bug report either! So we'll stick to doing it via the to_zarr method's 'encoding' argument. Basically throw away all the stuff before and brute force encode it in the .zmetadata file I guess. Funnily enough, this cuts the conversion time in half, from ~1 hour to ~30 min, so I'm happy with that!