:chart_with_upwards_trend: Rate of elev change over time up to 20200404
Calculate Antarctic rate of ice surface elevation change over time (dhdt) up to 20200404! Decided to run the dhdt calculation on data points with a height range of >0.25 m (n=~36 million), instead of just >0.5 m before (n=~10 million). More than 3 times the data points, and nice to see the time to do the calculation has just about tripled too from 3 min to 9 min. Not accounting for the fact that we're using float32 instead of float64, and the hours of hand holding to get it to work! Updated teaser image accordingly on main README.md too.