:children_crossing: Include referencegroundtrack via xarray preprocessor
Towards making transect/cross-section plots! Ensure that our ATL11 xarray dataset has the referencegroundtrack number in it, for easier tracking later on! Builds on f30110c79a3ad5dbdef628c150d45bf6febdbde9. I've updated the Kamb subset for my friend, and it's good to see that hvplot works interactively when zooming in now if datashade is on (but rasterize shows up blank?).

Ideally we would be using intake.open_ndzarr for this, but there are too many missing pieces in the puzzle to get it to work seamlessly now. So I've reverse engineered the add_path_to_ds mechanism and used it to extract the reference ground track number from the filename when loading from the Zarr files! The coordinates are stored as int64 which seems overkill (int8 would suffice), but that can be amended properly at a later date, hopefully upstream at intake/xarray.