:heavy_plus_sign: Add cuml and scikit-learn
GPU Machine Learning Algorithms, in Python! This also brings in cudf - a GPU DataFrame Library, cudatoolkit, and other RAPIDS AI libraries.

Pinned numba to 0.51.0 and bumped up llvmlite in the poetry.lock file to 0.34.0 to be in line with conda installed version and prevent "ERROR: Cannot uninstall 'llvmlite'. It is a distutils installed project and thus we cannot accurately determine which files belong to it which would lead to only a partial uninstall". This requires datashader 0.11.1 too, all in the fun of dependency management.