:sparkles: Ndarray to Parquet convenience function
Building on top of eb61ff609a7fa12a419cebb35daf45f072fb8ad3, but for n-dimensional arrays, and writing the dataframe to Parquet too! This function might be a little too convenient (read: contains hardcoding), but it smooths out some of the rough edges in terms of PyData file format interoperability. Should contribute this somewhere upstream when I get the time.