:alien: Update to poetry 1.1 build system and lock file format
As mentioned in https://python-poetry.org/blog/announcing-poetry-1-1-0.html, the build-system in the pyproject.toml has to be manually changed to use poetry-core>=1.0.0. Ran `poetry lock --no-update` to sort lots of stuff in the poetry.lock file to be in the order of name-version-description-category-etc as per poetry 1.1. Also had to explicitly pin python-snappy in pyproject.toml because poetry didn't install it via intake[server] for some reason.