:recycle: Better handling of file move and proj4 in spatiotemporal_cube
Refactor to move the NetCDF files into the folder correctly, overwriting the filepath(s) if the file is already present. Also added a 'projection' parameter so that the projection information is stored in the NetCDF grid and viewable directly in QGIS without needing to set a projection first. Using the proj4 string of Antarctic Polar Stereographic EPSG:3031 from https://spatialreference.org/ref/epsg/3031/proj4/. Note to self, the grid_mapping coordinate should actually be an attribute, either handle that using xarray.open_rasterio, use rioxarray, or wait until the pangeo/xarray community standardizes on what to do with this.