:hammer: Impose 3 median absolute deviation limit on grid surface output
Better handle elevation outliers that had resulted in sharp peaks and troughs in the output NetCDF grid. Doing so by telling `pygmt.surface` to limit the output solution to a lower and upper bound, determined as ±3 median absolute deviations from the median of the entire dataframe table. Originally I was clipping the grid in a post-processing step, but decided it's much better to impose the limits in the surface algorithm itself! The elevation statistics of the test case in Greenland also looks more sensible now than before.