:triangular_flag_on_post: Refresh crossover track plotting scripts
Main changes are an elev_filter setting in `deepicedrain.vizplots.plot_crossovers`, and the addition of annotated track numbers in the crossover_area map plot using GMT's quoted line functionality. Lots of other little bits too, including a refresh of the subglacial_lake ID (key) and name (value) pairs, a quick improvement to the inefficient brute-force crossover for-loop, and organizing figures inside 'figures/{lakename}' instead of just the 'figures' directory. Still a big chunk of local code yet to be committed, need to refactor and streamline the processing to be able to loop through many subglacial lakes in a more hands off manner!