:dizzy: 3D view of interpolated ice surface elev over ICESat-2 cycles
Visualizing ice surface changes in 3D, because it's more intuitive to the eye! There's a bit of an art to 3D perspective plots, and PyGMT's `grdview` does a fine job at it, though there's many lines of code involved! The figure consists of two plots. On the left is a shaded ice surface elevation grid overlaid with the subglacial lake outline in yellow. On the right is an elevation difference/anomaly grid view (relative to ICESat-2 Cycle 3).

Still need to wait for `subplot` and `plot3d` to be wrapped in PyGMT to simplify the code and make things look more polished. Animated GIF is made using imagemagick's `convert` in bash as it was higher quality (and less lines of code) than Pillow. Also have a code cell showing an alternative HvPlot-based 2D map of the same time-series grid, complete with an interactive slider to cycle through each ICESat-2 cycle.