:dizzy: Plot green ICESat-2 track points and lake outline on 3D surface
Now that `pygmt.plot3d` is available, we can plot data points and lines on top of our 3D grid surface at the correct elevation! The lake outline was plotted before at a single median height, now we plot the polygon nodes according to the grdtrack sampled z value and if it's a NaN, we use the median z value instead.