:chart_with_upwards_trend: 3D Ice surface elevation plotting function
Reduce cognitive discomfort by putting the very long 3D perspective plotting code into a tidy `plot_icesurface` function! This taking the mess of PyGMT code in atlxi_lake.ipynb and puts it in vizplots.py.  Inputs are a main grid and diff_grid, their bounding cube regions, optional track and lake polygon overlays, the 3D azimuth and elevation perspective, and a title/subtitle. Lots of hardcoded stuff but we can worry about those later.

A unit test was also added to ensure that 100+ lines of code will work. Lots of forseeable improvements such as use of proper subplots instead of `shift_origin` so that a single aligned title can be printed, and a convenience with-statement function that can produce 'Polar Stereographic (km)' labels. But maybe get an integration test in here first!