:sparkles: Active subglacial lake mega-cluster analysis
Look at not one but multiple active subglacial lakes! Examples include: Lake 78, Subglacial Lake Conway, Subglacial Lake Mercer and Subglacial Lake Whillans on the Siple Coast; and Recovery 2 on Recovery Glacier. Also renumbered all the lake ids for the single lakes. The subglacial_lakes.feature file splits the single and mega-cluster lake tests. Note that the mega-cluster test won't run on CI until deepicedrain v0.4.0 is released.

Really required a lot of refactoring to handle multiple lake polygons. Bounding box region is now determined using a convex hull so that it works for single or multiple polygons. Lakes are dissolved into a MultiPolygon (geopandas v0.8.0 requiring a dissolve field, problematic for Lake 78 that spans Whillans and Mercer basins). The lake outlines are saved to an OGR GMT format which is actually quite nice as it can be shared around.