:card_file_box: Shift subglacial lakename id pairs to atlas_catalog.yaml
Allow easier re-use of lake polygons across scripts by moving the messy Python dictionary of lake_ids and lake_name pairs that was in atlxi_lake.ipynb to the metadata section of atlas_catalog.yaml! Still scope for shortening that chunk of code I think, and definitely need to propagate this change to the subglacial_lakes.feature file once v0.4.0 is released.