:construction: Save ice volume displacement data to textfile
Like, actually calculate and save (!!) ice volume displacement (dvol), so that we can aggregrate the dvol values from several lakes into one plot later on, or share the data with other people. Both the nominal values and standard deviations are saved, though this requires the pint and uncertainties package. Also did some hacking to ensure pint unit registries are the same to avoid a "ValueError: Cannot operate with Quantity and Quantity of different registries."