:clown_face: Single 3D plot of mean ice surface elevation and dhdt trend
It's tough to animate on paper, so let's take the average and trend instead! The spatiotemporal_cube function will now calculate mean elevation (<z>) and rate of elevation change (dhdt) and include it in the output xarray.Dataset. The plot_icesurface code has been modified slightly to handle dhdt as well as dh, and we ensure the roma colormap is centered on zero. Getting all the ICESat-2 raw xyz points and one transect line was complicated (read: needs refactoring). Yellow transect is now plotted only on the new static 3D plot with dhdt trend, leaving the older 3D plots (which becomes a GIF) pretty much intact as before.