:technologist: Update buttons to mybinder.org and Pangeo BinderHub
Removing the broken Pangeo Binder link that is offline since Dec 2021, and replace it with the generic mybinder.org federation button. Also sneaking in the Pangeo BinderHub button created from nbgitpuller, which links to the AWS uswest-2 hosted instance. Fingers crossed that this allows for cloud access to to ICESat-2!